# backend/swagger.json
{
	"basePath": "/api",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"IOrder": {
			"properties": {
				"id": {
					"type": "string"
				},
				"name": {
					"type": "string"
				}
			},
			"required": [
				"id",
				"name"
			],
			"type": "object",
			"additionalProperties": false
		},
		"IUser": {
			"properties": {
				"email": {
					"type": "string"
				},
				"role": {
					"type": "string",
					"enum": [
						"ADMIN",
						"USER"
					],
					"x-nullable": true
				},
				"language": {
					"type": "string",
					"x-nullable": true
				},
				"password": {
					"type": "string"
				}
			},
			"required": [
				"email",
				"password"
			],
			"type": "object",
			"additionalProperties": false
		},
		"AuthCred": {
			"properties": {
				"email": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}
			},
			"required": [
				"email",
				"password"
			],
			"type": "object",
			"additionalProperties": false
		},
		"IUserExport": {
			"properties": {
				"email": {
					"type": "string"
				},
				"role": {
					"type": "string",
					"enum": [
						"ADMIN",
						"USER"
					],
					"x-nullable": true
				},
				"language": {
					"type": "string",
					"x-nullable": true
				}
			},
			"required": [
				"email"
			],
			"type": "object",
			"additionalProperties": false
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/order/{id}": {
			"get": {
				"operationId": "GetOne",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IOrder"
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IOrder"
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IOrder"
						}
					}
				]
			}
		},
		"/order": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IOrder"
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IOrder"
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IOrder"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IOrder"
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IOrder"
						}
					}
				]
			}
		},
		"/user/register": {
			"post": {
				"operationId": "CreateUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"properties": {
								"token": {
									"type": "string"
								}
							},
							"required": [
								"token"
							],
							"type": "object"
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IUser"
						}
					}
				]
			}
		},
		"/user/login": {
			"post": {
				"operationId": "LoginUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"properties": {
								"token": {
									"type": "string"
								}
							},
							"required": [
								"token"
							],
							"type": "object"
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/AuthCred"
						}
					}
				]
			}
		},
		"/user/userInfo": {
			"get": {
				"operationId": "GetUserByToken",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IUserExport"
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "token",
						"required": true,
						"type": "string"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "token",
			"description": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFkbWluQGFkbWluIiwiaWQiOjgsInJvbGUiOiJBRE1JTiIsImlhdCI6MTYyNzI5MjAyOX0.smA9a5v47rs-oG2loD1jR5Zd5RlGqmb0XN0KWjniyKo",
			"in": "header"
		}
	},
	"tags": [
		{
			"name": "Products",
			"description": "Операции доступные с продуктами"
		},
		{
			"name": "User",
			"description": "Операции с пользователем"
		}
	]
}